Finished groundwork for placing pipes.
Right now there's no graphical interface for selecting pipes, but all the controls are there (not including switching pipe types, but this works if the starting type is changed).
From here, a solution check can be implemented, as well as finishing the presentation of each level (centering the level map and adding a background) and adding an exit entity.